Guard Mutations, various bugfixes and code cleanup
Implementation of guard mutations handling and variablizations, though
it still wont work until mujava gives me the correct mutatedLine where
the mutgenlimit is located.

Using PRVO refined and configuring it not to generate super mutations.

Fixing while unroll comments persist through unroll to ifs and cleanup
of the code.

Disabling extra check of duplicates, we trust the hash digest ;)

Fix of feedback apply given when after feedback a mutation didnt compile
or was duplicate, the next mutator was computed using the feedback and
that was skipping mutants.

General code cleanup