Fix in incorrect pruning of uncompilable mutations
It was assuming a line where there was an error, if this line was
mutable, that the current mutation index at that line had to be ignored
thereafter.
But if the error was an 'unreachable code' one, the error could've been
introduced by a mutation in another line, so now when the error is an
'unreachable code' one, the mutation at the error line is not marked to
be skipped in the future.